Address infinite loops while reading ID3v2 tags. (#48)
The `Buffer.read()` call will infinitely loop if the number of bytes
remaining in the file is less than the requested size. A large number of
`read()` calls in `ID3v2Parser` are at risk of infinitely looping on
malformed files. I'm changing `Buffer` to throw an exception when it
detects there are no more bytes to read from the file.

There is a more egregious case with the potential `Xing` header. The
parser blindly reads 1500 bytes to search for the header, but there is
no guarantee (even in a well-formed file) that there will be 1500 more
bytes remaining to be read. I'm introducing a `readAtMost` method to
`Buffer` which will handle cases where there are fewer than `size` bytes
remaining in the file.

I'm also updating some of the logic which is checking for the start of
the `Xing` header to be safer, checking to ensure that there are enough
bytes remaining to parse the data from this header.

Co-authored-by: Loren Van Spronsen <lorenvs@google.com>